upload lambda: disable for image based lambdas
Problem: The toolkit only supports 'Upload Lambda' when lambdas have a
package type of 'zip'. Attempting to upload to an image based lambda results in an error.

Solution: Continue to show but disable the command on lambda nodes that
have a package type of 'image'. The lambdas listed in the
UploadLambdaWizard with this package type will be invalid (the user will
not be able to select it) and a decription has been added.